Make the editor fall under the route of graph-editor
