+++ src/routes/auth/signout/+page.server.ts
