+++ src/routes/graph-editor/+layout.server.ts
@@ -1,3 +1,4 @@
+import prisma from '$lib/server/db/prisma.js';
 import { redirect } from '@sveltejs/kit';
 
 // This is to test the connection to the database
@@ -6,9 +7,7 @@ import { redirect } from '@sveltejs/kit';
 export const load = async ({ locals, url }) => {
 	const session = await locals.auth();
 
-	if (!session?.user) throw redirect(303, '/auth');
-
 	return {
-		session
+		testConnection: prisma.program.findFirst()
 	};
 };
