+++ src/routes/graph-editor/SearchCourses.svelte
