+++ src/routes/graph-editor/courses/+page.svelte
