+++ src/routes/graph-editor/courses/[code]/+page.svelte
