+++ src/routes/graph-editor/courses/[code]/CreateNewGraphButton.svelte
@@ -5,21 +5,23 @@
 	import { toast } from 'svelte-sonner';
 	import { type Infer, superForm, type SuperValidated } from 'sveltekit-superforms';
 	import { zodClient } from 'sveltekit-superforms/adapters';
-	import { programSchema } from './zodSchema';
+	import { graphSchema } from '$lib/zod/graphSchema';
+	import type { Course } from '@prisma/client';
 
 	type Props = {
-		form: SuperValidated<Infer<typeof programSchema>>;
+		form: SuperValidated<Infer<typeof graphSchema>>;
+		course: Course;
 	};
 
-	const { form: programForm }: Props = $props();
+	const { form: graphForm, course }: Props = $props();
 
 	let dialogOpen = $state(false);
 
-	const form = superForm(programForm, {
-		validators: zodClient(programSchema),
+	const form = superForm(graphForm, {
+		validators: zodClient(graphSchema),
 		onResult: ({ result }) => {
 			if (result.type == 'success') {
-				toast.success('Program created successfully!');
+				toast.success('Graph created successfully!');
 				dialogOpen = false;
 			}
 		}
@@ -31,22 +33,24 @@
 <DialogButton
 	bind:open={dialogOpen}
 	icon="plus"
-	button="New Program"
-	title="Create Program"
-	description="Programs are collections of Courses, usually pertaining to the same field of study. Looking to try
-	out the Graph editor? Try making a sandbox environment instead!"
+	button="New Graph"
+	title="Create Graph"
+	description="Graphs are collections of Nodes and Edges, usually pertaining to the same field of study. Create this new graph in the {course.name}"
+	class="col-span-2"
 >
 	<!-- For sumbitting a NEW PROGRAM
  	It triggers an action that can be seen in +page.server.ts -->
-	<form action="?/new-program" method="POST" use:enhance>
+	<form action="?/add-graph-to-course" method="POST" use:enhance>
+		<input type="hidden" name="courseCode" value={course.code} />
+
 		<Form.Field {form} name="name">
 			<Form.Control>
 				{#snippet children({ props })}
-					<Form.Label for="name">Program name</Form.Label>
+					<Form.Label for="name">Graph name</Form.Label>
 					<Input {...props} bind:value={$formData['name']} />
 				{/snippet}
 			</Form.Control>
-			<Form.Description />
+			<Form.Description>A common name for the graph</Form.Description>
 			<Form.FieldErrors />
 		</Form.Field>
