+++ src/routes/graph-editor/courses/[code]/graphs/[graphid]/+layout.svelte
