+++ src/routes/graph-editor/courses/[code]/graphs/[graphid]/Preview.svelte
