+++ src/routes/graph-editor/courses/[code]/graphs/[graphid]/domains/CreateNewDomainRel.svelte
