+++ src/routes/graph-editor/courses/[code]/graphs/[graphid]/domains/change-color/+server.ts
