+++ src/routes/graph-editor/courses/[code]/graphs/[graphid]/lectures/+page.svelte
