+++ src/routes/graph-editor/courses/[code]/graphs/[graphid]/subjects/+page.server.ts
