+++ src/routes/graph-editor/courses/[code]/graphs/[graphid]/subjects/ChangeSubject.svelte
