+++ src/routes/graph-editor/courses/[code]/graphs/[graphid]/subjects/CreateNewSubject.svelte
