+++ src/routes/graph-editor/courses/[code]/graphs/[graphid]/subjects/DeleteSubject.svelte
