+++ src/routes/graph-editor/programs/[uuid]/settings/schema.ts
